backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
APScheduler
passlib[bcrypt]
PyJWT
python-dotenv
